+++ include/users/modules/class-bulk-delete-users-by-user-meta.php
@@ -110,13 +110,7 @@ class Bulk_Delete_Users_By_User_Meta extends BD_User_Meta_Box_Module {
 		$delete_options['meta_compare']   = array_get( $_POST, 'smbd_u_meta_compare', '=' );
 		$delete_options['meta_value']     = array_get( $_POST, 'smbd_u_meta_value' );
 
-		$delete_options['login_restrict'] = array_get_bool( $_POST, "smbd_{$this->field_slug}_login_restrict", false );
-		$delete_options['login_days']     = absint( array_get( $_POST, "smbd_{$this->field_slug}_login_days", 0 ) );
-
-		$delete_options['no_posts']       = array_get_bool( $_POST, "smbd_{$this->field_slug}_no_posts", false );
-		$delete_options['limit_to']       = absint( array_get( $_POST, "smbd_{$this->field_slug}_limit_to", 0 ) );
-
-		$this->process_delete( $delete_options );
+		$this->process_user_delete( $delete_options );
 	}
 
 	/**